fix : 공지사항 수정 API 수정
수정 내용
- 사진 없이 등록한 공지사항은 사진을 등록하려고 할때 저장이안되는 오류가 발생해 코드수정

+++ src/main/java/com/team/RecipeRadar/domain/notice/application/NoticeServiceImpl.java
@@ -102,15 +102,18 @@ public class NoticeServiceImpl implements NoticeService {
         if(!notice.getMember().getLoginId().equals(loginId)) throw new AccessDeniedException("관리자만 삭제 가능합니다.");
 
         UploadFile uploadFile = imgRepository.getOriginalFileName(notice.getId());
-        if(file!=null) {
+        if(file!=null && uploadFile!=null) {
             if (!uploadFile.getOriginFileName().equals(file.getOriginalFilename())) {       // 원본파일명이 다를경우에만 s3에 기존 사진을 삭제 후 새롭게 저장
                 s3UploadService.deleteFile(uploadFile.getStoreFileName());
                 String storedFileName = s3UploadService.uploadFile(file);
                 uploadFile.update(storedFileName, file.getOriginalFilename());
                 imgRepository.save(uploadFile);
             }
+        }else if(file!=null && uploadFile==null){// 처음부터 파일을 저장하지 않았을때 저장
+            String uploadedFile = s3UploadService.uploadFile(file);
+            UploadFile new_notice = UploadFile.builder().originFileName(file.getOriginalFilename()).storeFileName(uploadedFile).notice(notice).build();
+            imgRepository.save(new_notice);
         }
-
         notice.update(adminUpdateRequest.getNoticeTitle(), adminUpdateRequest.getNoticeContent());
         noticeRepository.save(notice);
     }
